Resume Challenge Project-1
This Power BI project is part of the Codebasics Resume Challenge and focuses on solving a real-world business case for AtliQ Grands, a luxury hotel chain in India. The goal is to analyze historical booking data and deliver actionable insights through an interactive dashboard. The project includes data modeling, DAX calculations, and visual storytelling to support strategic decision-making and revenue recovery.